Add wlr_texture_set
When performing cross-GPU operations we need to keep track of where
a buffer has been imported and where it has been copied. This is
what a texture set does. It keeps a dictionary of renderers and
if a wlr_texture has been created for them. The user can request a
texture for a particular renderer, and it will perform any blits
needed and cache the result.