backend/x11: re-send last buffer on expose event
Instead of requesting a new frame from the compositor when we
receive an expose event, re-submit the last frame. This makes the
X11 backend behave like the other backends and removes the last
wlr_output_update_needs_frame() call in backends, at the cost of
keeping the current buffer locked for potentially a longer time.