fifo-v1: new protocol implementation
Implement fifo-v1 protocol: the capability for clients to explicitly queue wl_surface
commits.

The design of the wl_surface interface defines a double buffer, which is intended to support only
(from the point of view of the client) what is sometimes called the 'mailbox' presentation mode:
each request on a wl_surface operates on an accumulated pending state until a .commit request is
issued, immediately and atomically replacing the wl_surface's current state. This means that only
the state related to the latest .commit request will be presented.

fifo-v1 is a protocol that aims to support an important use case, namely the FIFO display
presentation mode, which allows to save the accumulated pending state for future presentations;
instead of a double buffer, a queue is used where each .commit request puts the accumulated pending
state.
This protocol effectively changes the behaviour of the .commit request on a wl_surface.

Signed-off-by: Sergio Gómez <sergio.g.delreal@gmail.com>