render: open platform device render node
Right now we rely on Mesa's renderonly/kmsro to handle split
render/display systems. And we have a bunch of special cases in
wlroots to accomodate for the fact we don't know about the render
node.

However, after a recent discussion [1], it appears we should be
able to handle split render/display systems ourselves. If the
backend is using a platform display-only device, we can pick a
platform render-only device. This should work for all known split
render/display SoCs.

[1]: https://gitlab.freedesktop.org/mesa/mesa/-/issues/9920