backend/drm: Ensure plane rotation is always reset
The rotation property of a plane persists between commits.
If any application has set the property before wlroots takes
over, then that rotation will still apply to wlroots' output.
If the rotate applied to a primary plane but not the cursor
plane (as would happen if the kernel command line contains
"video=HDMI-A-1:1920x1080,rotate=180"), then system operation
becomes "interesting".

If the property exists, always set it to DRM_MODE_ROTATE_0
to ensure that the system generates the expected output.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>