backend/headless: inform about the refresh period in the wlr_output 'present' event
Upcoming commit_timing_v1 needs precise wlr_output timing information to schedule future
presentations.
We are going to add the features to the headless backend to be able to simulate the timings of a
real output so that compositors which use it can synchronize with their external timings (e.g,
gamescope get their output timings outside wlroots, but would like to use commit_timing_v1 through
the headless backend).
This commit is the first step towards that by tracking the exact refresh period (in nanoseconds) and
inform about it through the wlr_output 'present' event.

Signed-off-by: Sergio Gómez <sergio.g.delreal@gmail.com>