editorconfig: Unset maximum line length for .git
An editorconfig's asterix section applies to the whole file tree,
including the .git folder. The Git commit message dialog is a temporary
file named .git/COMMIT_EDITMSG, which the editorconfig applies to.

To avoid having the line length of commit messages changed to 100 chars,
add a rule to specifically unset the max_line_length for the .git
folder.