util/box: use nextafter(3) in wlr_box_closest_point()
Given two boxes placed next to each other in the integer coordinate
space, previously they would've been considered to be disjoint in the
float64 coordinate space by wlr_box_closest_point(), as if there was a
1-unit gap. Practically this meant inability to move a cursor between
outputs if the modulus of the cursor position delta was less than 1.0.

This commit ensures that there is no point representable as a pair of
float64 numbers less than (x + width, y + height) respectively while
considered to be outside the box by wlr_box_closest_point().

Fixes: 3fc66d4525916b9301236a000a6ed03311ed25a7