render/vulkan: Skip staging buffer
Staging buffers can be more performant for dGPUs, but not necessarily
for APUs. radeonsi has a heuristic that migrates the target texture to
be LINEAR after 10 transfers, which also disables the use of staging
texture.

A more advanced solution would swap between the two as needed, but for
now stick to not using a staging buffer.

This commit by itself slightly worsens performance in my test, but
allows us to reasily implement reuse of mapping in a follow-up commit
which negates the hit.